backend/login_service/requirements.txt
flask
flask_sqlalchemy
mysql-connector-python
werkzeug

flask_bcrypt


pyjwt

flask_cors

flask_mail